+++ app/scripts/comp/launcher/launcher-cordova.js
@@ -208,9 +208,6 @@ const Launcher = {
     minimizeApp() {
         this.hideApp();
     },
-    canMinimize() {
-        return false;
-    },
     canDetectOsSleep() {
         return false;
     },