+++ release-notes.md
@@ -4,6 +4,7 @@ Release notes
 `+` YubiKey integration in two modes: OATH and Challenge-Response  
 `+` configs are now encrypted with a key stored in keychain  
 `+` #1460: auto-type on double-clicking field labels  
+`+` #830: minimize instead of close option on macOS  
 `+` #557: Argon2 speed improvements in desktop apps  
 `+` #1503: ARM64 Windows support  
 `+` #1480: option to create a portable installation