+++ release-notes.md
@@ -1,6 +1,6 @@
 Release notes
 -------------
-##### v1.15.0 (WIP)
+##### v1.15.0 (2020-06-06)
 `+` YubiKey integration in two modes: OATH and Challenge-Response  
 `+` configs are now encrypted with a key stored in keychain  
 `*` default format for new files is now KDBX4